Added output to xlsx
added output to xlsx with each value in a different column so they can
be viewed more easily. To speed up the processing, all data is only
written to the console if the verbose flag -v is set. Script can now
also be used on non-unix file systems such as Windows and can be used
with no arguments given.